Refactor #PP-49: 멤버 엔티티 통합
